File structure overhaul [Tested]
